# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkGuard — Telegram Mini App for link safety checking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  safe: "safe",
  caution: "caution",
  danger: "danger",
  unknown: "unknown",
} as const;

export interface CheckLinkBody {
  url: string;
  telegramUserId: string;
}

export interface LinkCheckResult {
  id: number;
  url: string;
  normalizedUrl: string;
  verdict: Verdict;
  threatTypes: string[];
  explanation: string;
  checkedAt: string;
  isTrustedDomain: boolean;
}

export interface LinkHistoryItem {
  id: number;
  url: string;
  verdict: Verdict;
  explanation: string;
  checkedAt: string;
  reportedToParent: boolean;
}

export interface LinkHistoryResponse {
  items: LinkHistoryItem[];
  total: number;
}

export interface LinkStatsResponse {
  totalChecked: number;
  safeCount: number;
  cautionCount: number;
  dangerCount: number;
  unknownCount: number;
  reportsToParent: number;
}

export interface ReportToParentBody {
  linkCheckId: number;
  telegramUserId: string;
}

export interface ReportToParentResponse {
  success: boolean;
  message: string;
}

export interface UserSettings {
  telegramUserId: string;
  parentContact?: string | null;
  trustedDomains: string[];
  notifyParentOnDanger: boolean;
  language: string;
}

export interface UpdateUserSettingsBody {
  telegramUserId: string;
  parentContact?: string | null;
  trustedDomains?: string[];
  notifyParentOnDanger?: boolean;
}

export interface RegisterUserBody {
  telegramUserId: string;
  firstName: string;
  lastName?: string | null;
  username?: string | null;
}

export interface UserProfile {
  telegramUserId: string;
  firstName: string;
  lastName?: string | null;
  username?: string | null;
  createdAt: string;
}

export type GetLinkHistoryParams = {
  telegramUserId: string;
  limit?: number;
};

export type GetLinkStatsParams = {
  telegramUserId: string;
};

export type GetUserSettingsParams = {
  telegramUserId: string;
};