# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkGuard — Telegram Mini App for link safety checking
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Check a URL for safety
 */
export const CheckLinkBody = zod.object({
  url: zod.string(),
  telegramUserId: zod.string(),
});

export const CheckLinkResponse = zod.object({
  id: zod.number(),
  url: zod.string(),
  normalizedUrl: zod.string(),
  verdict: zod.enum(["safe", "caution", "danger", "unknown"]),
  threatTypes: zod.array(zod.string()),
  explanation: zod.string(),
  checkedAt: zod.coerce.date(),
  isTrustedDomain: zod.boolean(),
});

/**
 * @summary Get link check history for current user
 */
export const getLinkHistoryQueryLimitDefault = 20;

export const GetLinkHistoryQueryParams = zod.object({
  telegramUserId: zod.coerce.string(),
  limit: zod.coerce.number().default(getLinkHistoryQueryLimitDefault),
});

export const GetLinkHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      url: zod.string(),
      verdict: zod.enum(["safe", "caution", "danger", "unknown"]),
      explanation: zod.string(),
      checkedAt: zod.coerce.date(),
      reportedToParent: zod.boolean(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get aggregated stats for a user
 */
export const GetLinkStatsQueryParams = zod.object({
  telegramUserId: zod.coerce.string(),
});

export const GetLinkStatsResponse = zod.object({
  totalChecked: zod.number(),
  safeCount: zod.number(),
  cautionCount: zod.number(),
  dangerCount: zod.number(),
  unknownCount: zod.number(),
  reportsToParent: zod.number(),
});

/**
 * @summary Send a link check result to the parent contact
 */
export const ReportToParentBody = zod.object({
  linkCheckId: zod.number(),
  telegramUserId: zod.string(),
});

export const ReportToParentResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get user settings
 */
export const GetUserSettingsQueryParams = zod.object({
  telegramUserId: zod.coerce.string(),
});

export const getUserSettingsResponseLanguageDefault = `ru`;

export const GetUserSettingsResponse = zod.object({
  telegramUserId: zod.string(),
  parentContact: zod.string().nullish(),
  trustedDomains: zod.array(zod.string()),
  notifyParentOnDanger: zod.boolean(),
  language: zod.string().default(getUserSettingsResponseLanguageDefault),
});

/**
 * @summary Update user settings
 */
export const UpdateUserSettingsBody = zod.object({
  telegramUserId: zod.string(),
  parentContact: zod.string().nullish(),
  trustedDomains: zod.array(zod.string()).optional(),
  notifyParentOnDanger: zod.boolean().optional(),
});

export const updateUserSettingsResponseLanguageDefault = `ru`;

export const UpdateUserSettingsResponse = zod.object({
  telegramUserId: zod.string(),
  parentContact: zod.string().nullish(),
  trustedDomains: zod.array(zod.string()),
  notifyParentOnDanger: zod.boolean(),
  language: zod.string().default(updateUserSettingsResponseLanguageDefault),
});

/**
 * @summary Register or update a Telegram user
 */
export const RegisterUserBody = zod.object({
  telegramUserId: zod.string(),
  firstName: zod.string(),
  lastName: zod.string().nullish(),
  username: zod.string().nullish(),
});

export const RegisterUserResponse = zod.object({
  telegramUserId: zod.string(),
  firstName: zod.string(),
  lastName: zod.string().nullish(),
  username: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});